% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bespoke_class-fit.R
\name{.bespoke_class_bridge}
\alias{.bespoke_class_bridge}
\title{Create the "Model"}
\usage{
.bespoke_class_bridge(processed, fn, ...)
}
\arguments{
\item{processed}{Processed inputs molded by hardhat.}

\item{fn}{A function that will generate "predictions" from predictors.}

\item{...}{Additional parameters passed on to \code{fn}.}
}
\value{
A \code{bespoke_class} model object.
}
\description{
This bridge is the same for each path into bespoke_class. It takes the
processed predictors (which are now a tibble) and outcome (which is now a
factor) and, in this case, makes sure they make sense with the supplied
function.
}
\keyword{internal}
