% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bespoke_classification-fit.R
\name{.bespoke_classification_impl}
\alias{.bespoke_classification_impl}
\title{Validate the Model Function}
\usage{
.bespoke_classification_impl(predictors, outcome, fn, ...)
}
\arguments{
\item{predictors}{A tibble of predictors.}

\item{outcome}{A factor of output classes.}

\item{fn}{A function that will generate "predictions" from predictors.}

\item{...}{Additional parameters passed on to \code{fn}.}
}
\value{
A list with elements \code{fn}, \code{dots}, and \code{outcome_levels}.
}
\description{
Normally this would be the place where the actual fit happens. Here, we
simply validate that the function makes sense with the input data.
}
\keyword{internal}
