% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bespoke_class-predict.R
\name{.factorize_predictions}
\alias{.factorize_predictions}
\alias{.factorize_predictions.default}
\alias{.factorize_predictions.integer}
\alias{.factorize_predictions.factor}
\alias{.factorize_predictions.character}
\title{Coerce Predictions to Factor}
\usage{
.factorize_predictions(predictions, outcome_levels)

\method{.factorize_predictions}{default}(predictions, outcome_levels)

\method{.factorize_predictions}{integer}(predictions, outcome_levels)

\method{.factorize_predictions}{factor}(predictions, outcome_levels)

\method{.factorize_predictions}{character}(predictions, outcome_levels)
}
\arguments{
\item{predictions}{Predictions produced by a bespoke model function.}

\item{outcome_levels}{A character vector indicating the expected outcome
levels, in the same order that they were defined in the training data.}
}
\value{
Predictions as a factor with the expected levels.
}
\description{
Coerce Predictions to Factor
}
\keyword{internal}
