% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bespoke_classification-predict.R
\name{.get_bespoke_classification_predict_function}
\alias{.get_bespoke_classification_predict_function}
\title{Choose the Predict Function for a Bespoke Classification}
\usage{
.get_bespoke_classification_predict_function(type)
}
\arguments{
\item{type}{A single character. The type of predictions to generate.
Valid options are:
\itemize{
\item \code{"class"} for "hard" class predictions.
\item \code{"prob"} for class probabilities.
}}
}
\value{
The predict function.
}
\description{
This is pointless right now but I'm keeping it in place to make it relatively
easy to implement probabilities.
}
\keyword{internal}
