% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bespoke_class-constructor.R
\name{.new_bespoke_class}
\alias{.new_bespoke_class}
\title{Construct a Bespoke Classification Model Object}
\usage{
.new_bespoke_class(fn, dots, outcome_levels, blueprint)
}
\arguments{
\item{fn}{A user-supplied function.}

\item{dots}{Additional parameters to pass onto the function.}

\item{outcome_levels}{The levels that were present in the training data.}

\item{blueprint}{The \code{{hardhat}} blueprint to prepare the predictors.}
}
\value{
A \code{\link[hardhat:new_model]{hardhat::new_model()}}.
}
\description{
Construct a Bespoke Classification Model Object
}
\keyword{internal}
