% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bespoke_class-predict.R
\name{.predict_bespoke_class_bridge}
\alias{.predict_bespoke_class_bridge}
\title{Prepare bespoke_class Data for Prediction}
\usage{
.predict_bespoke_class_bridge(type, model, predictors)
}
\arguments{
\item{type}{A single character. The type of predictions to generate.
Valid options are:
\itemize{
\item \code{"class"} for "hard" class predictions.
\item \code{"prob"} for class probabilities. (not yet implemented)
}}

\item{model}{A \code{bespoke_class} model object. hardhat switches to model from
here forward, I think because the object has to actually be a model at this
point for anything to make sense, and they don't have to fight the predict
generic for the name.}

\item{predictors}{Forged predictor data.}
}
\value{
A tibble with output dependent on type.
}
\description{
Prepare bespoke_class Data for Prediction
}
\keyword{internal}
