% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bespoke_class-predict.R
\name{.predict_bespoke_class_prob}
\alias{.predict_bespoke_class_prob}
\title{Predict Probabilities for Bespoke Classification Models}
\usage{
.predict_bespoke_class_prob(model, predictors)
}
\arguments{
\item{model}{A \code{bespoke_class} model object. hardhat switches to model from
here forward, I think because the object has to actually be a model at this
point for anything to make sense, and they don't have to fight the predict
generic for the name.}

\item{predictors}{Forged predictor data.}
}
\value{
A tibble with class probabilities (1 or 0).
}
\description{
Predict Probabilities for Bespoke Classification Models
}
\keyword{internal}
