% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bespoke_classification-predict.R
\name{.predict_bespoke_classification_class}
\alias{.predict_bespoke_classification_class}
\title{Predict Classes for Bespoke Classification Models}
\usage{
.predict_bespoke_classification_class(model, predictors)
}
\arguments{
\item{model}{A \code{bespoke_classification} model object. hardhat switches to
model from here forward, I think because the object has to actually be a
model at this point for anything to make sense, and they don't have to
fight the predict generic for the name.}

\item{predictors}{Forged predictor data.}
}
\value{
A tibble with a factor column identifying the outcome.
}
\description{
Predict Classes for Bespoke Classification Models
}
\keyword{internal}
