% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bespoke_classification-predict.R
\name{predict.bespoke_classification}
\alias{predict.bespoke_classification}
\title{Predict from a \code{bespoke_classification}}
\usage{
\method{predict}{bespoke_classification}(object, new_data, type = c("class", "prob"), ...)
}
\arguments{
\item{object}{A \code{bespoke_classification} object.}

\item{new_data}{A data frame or matrix of new predictors.}

\item{type}{A single character. The type of predictions to generate.
Valid options are:
\itemize{
\item \code{"class"} for "hard" class predictions.
\item \code{"prob"} for class probabilities.
}}

\item{...}{Not used, but required for extensibility.}
}
\value{
A tibble of predictions. The number of rows in the tibble is guaranteed
to be the same as the number of rows in \code{new_data}.
}
\description{
Predict from a \code{bespoke_classification}
}
