% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bespoke-package.R
\docType{package}
\name{bespoke-package}
\alias{bespoke}
\alias{bespoke-package}
\title{bespoke: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
